+++ src/sentry/native/native_sdk.h
@@ -36,7 +36,6 @@ public:
 	virtual String capture_event(const Ref<SentryEvent> &p_event) override;
 
 	virtual void initialize() override;
-	virtual bool is_initialized() override { return initialized; }
 
 	virtual ~NativeSDK() override;
 };